task and labels relations at stage 6 trial 1

+++ server/routes/tasks.js
@@ -1,7 +1,7 @@
 // @ts-check
 
 import i18next from 'i18next';
-// import _ from 'lodash';
+import _ from 'lodash';
 
 export default (app) => {
   app
@@ -32,8 +32,51 @@ export default (app) => {
       });
       return reply;
     })
-    .post('/tasks', { name: 'tasks#create', preValidation: app.authenticate }, async (req, reply) => {
-      const task = new app.objection.models.task();
+    .post(
+      '/tasks',
+      { name: 'tasks#create', preValidation: app.authenticate },
+      async (req, reply) => {
+        const creatorId = req.user.id;
+        const {
+          name, description, statusId, executorId,
+        } = req.body.data;
+        const taskData = {
+          name,
+          description,
+          creatorId,
+          statusId: Number(statusId),
+        };
+        // eslint-disable-next-line no-unused-expressions
+        executorId !== '' ? taskData.executorId = Number(executorId) : null;
+        const labelsIds = _.toArray(_.get(req.body.data, 'labels', []));
+        const thisTaskLabels = await app.objection.models.taskLabel
+          .query()
+          .findByIds(labelsIds);
+        taskData.labels = thisTaskLabels;
+        try {
+          await app.objection.models.task.transaction(async (trx) => {
+            await app.objection.models.task
+              .query(trx)
+              .insertGraph(taskData, { relate: true });
+          });
+          req.flash('info', i18next.t('flash.tasks.create.success'));
+          return reply.redirect(app.reverse('tasks#index'));
+        } catch ({ data: errors }) {
+          const taskStatuses = await app.objection.models.taskStatus.query().debug();
+          const users = await app.objection.models.user.query();
+          const taskLabels = await app.objection.models.taskLabel.query().debug();
+          req.flash('error', i18next.t('flash.tasks.create.error'));
+          reply.render('tasks/new', {
+            task: taskData, users, taskStatuses, taskLabels, errors,
+          });
+          return reply;
+        }
+      },
+    )
+    .patch('/tasks/:id/', { name: 'tasks#update', preValidation: app.authenticate }, async (req, reply) => {
+      const taskId = req.params.id;
+      const task = await app.objection.models.task.query().withGraphJoined('[status, creator, executor, labels]')
+        .findById(taskId);
       const creatorId = req.user.id;
       const {
         name, description, statusId, executorId,
@@ -46,62 +89,43 @@ export default (app) => {
       };
       // eslint-disable-next-line no-unused-expressions
       executorId !== '' ? taskData.executorId = Number(executorId) : null;
-      taskData.labels = req.body.data.labels;
-      reply.send(taskData);
-      // task.$set(taskData);
-      // try {
-      //   const validtask = await app.objection.models.task
-      //     .fromJson(taskData);
-      //   await app.objection.models.task.query().insert(validtask);
-      //   req.flash('success', i18next.t('flash.tasks.create.success'));
-      //   reply.redirect(app.reverse('tasks#index'));
-      //   return reply;
-      // } catch ({ data }) {
-      //   const taskStatuses = await app.objection.models.taskStatus.query().debug();
-      //   const users = await app.objection.models.user.query();
-      //   req.flash('error', i18next.t('flash.tasks.new.error'));
-      //   return reply.render('tasks/new', {
-      //     task, taskStatuses, users, errors: data,
-      //   });
-      // }
-    })
-    .patch('/tasks/:id/', { name: 'tasks#update', preValidation: app.authenticate }, async (req, reply) => {
-      const taskId = req.params.id;
-      const {
-        name, description, statusId, executorId,
-      } = req.body.data;
-      const taskData = {
-        name,
-        description,
-        statusId: Number(statusId),
-      };
-      // eslint-disable-next-line no-unused-expressions
-      executorId !== '' ? taskData.executorId = Number(executorId) : taskData.executorId = null;
-      const task = await app.objection.models.task.query().findById(taskId);
-      task.$set(taskData);
+      const labelsIds = _.toArray(_.get(req.body.data, 'labels', []));
+      const thisTaskLabels = await app.objection.models.taskLabel.query().findByIds(labelsIds);
       try {
-        await task.$query().patch(taskData);
+        await app.objection.models.task.transaction(async (trx) => {
+          await task.$relatedQuery('labels', trx).unrelate();
+          await task.$query(trx).patch(taskData);
+          await Promise.all(thisTaskLabels.map((label) => task.$relatedQuery('labels', trx).relate(label)));
+        });
         req.flash('success', i18next.t('flash.tasks.edit.success'));
         return reply.redirect(app.reverse('tasks#index'));
       } catch ({ data: errors }) {
         const taskStatuses = await app.objection.models.taskStatus.query().debug();
         const users = await app.objection.models.user.query();
+        const taskLabels = await app.objection.models.taskLabel.query().debug();
+        task.$set(taskData);
         req.flash('error', i18next.t('flash.tasks.edit.error'));
         reply.render('tasks/edit', {
-          task, taskStatuses, users, errors,
+          task, taskStatuses, taskLabels, users, errors,
         });
         return reply;
       }
     })
     .delete('/tasks/:id', { name: 'tasks#destroy', preValidation: app.authenticate }, async (req, reply) => {
+      const currentUserId = req.user.id;
       const taskId = req.params.id;
+      const task = await app.objection.models.task.query().findById(taskId);
+      if (currentUserId !== task.creatorId) {
+        req.flash('error', i18next.t('flash.tasks.delete.accessError'));
+        return reply.redirect(app.reverse('tasks#index'));
+      }
       await app.objection.models.task.query().deleteById(taskId);
       req.flash('success', i18next.t('flash.tasks.delete.success'));
       return reply.redirect(app.reverse('tasks#index'));
     })
     .get('/tasks/:id', { name: 'tasks#show', preValidation: app.authenticate }, async (req, reply) => {
       const task = await
-      app.objection.models.task.query().withGraphJoined('[status, creator, executor]').findById(req.params.id).debug();
+      app.objection.models.task.query().withGraphJoined('[status, creator, executor, labels]').findById(req.params.id).debug();
       reply.render('tasks/singleTask', { task });
       return reply;
     });
